requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
gunicorn
Pillow
django-extensions
django-crontab
